created wishist object
