+++ pages/index.js
@@ -2,38 +2,25 @@ import { useState, useEffect } from "react";
 import styled from "styled-components";
 import Post from "../component/Post";
 
-export default function Home() {
-  const [postList, setPostList] = useState([]);
-  const [wholeList, setWholeList] = useState([]);
+export default function Home({ postList }) {
+  const [filteredPostList, setFilteredPostList] = useState([]);
   const [searchText, setSearchText] = useState("");
 
+  // set filteredPostList with postList value
+  //postList is a list of all posts fetched from db
   useEffect(() => {
-    const fetchData = async () => {
-      const data = await fetch("/api");
-      const dataResponse = await data.json();
-
-      setWholeList(dataResponse);
-      setPostList(dataResponse);
-    };
-    fetchData().catch(console.error);
-  }, []);
+    setFilteredPostList(postList);
+  }, [postList]);
 
   function handleSearch(event) {
     const word = event.target.value;
     setSearchText(word);
-    const searchResult = wholeList.filter((post) => {
+    const searchResult = postList.filter((post) => {
       const postName = post.name.toLowerCase();
       return postName.includes(word);
     });
-    setPostList(searchResult);
+    setFilteredPostList(searchResult);
   }
-
-  // async function handleReload(event) {
-  //   const form = event.target;
-  //   const data = await fetch("/api/posts");
-  //   const postData = await data.json();
-  //   setPostList(postData);
-  // }
 
   return (
     <>
@@ -48,9 +35,9 @@ export default function Home() {
             onChange={handleSearch}
           ></Search>
         </SearchBox>
-        {/* <button onClick={handleReload}>Refresh</button> */}
+
         <PostList>
-          {postList.map((post) => {
+          {filteredPostList.map((post) => {
             return <Post postData={post} key={post._id}></Post>;
           })}
         </PostList>
